zshrc: alias k should work with kubectl exec
The previous k function doesn't work in situations like:

```console
$ k exec -it <pod> -c <container> -- sh
```

because it is expanded into:

```console
$ kubectl exec -it <pod> -c <container> -- sh --context=<context>
```

It doesn't choose the correct context, and `--context=<context>` isn't a valid
parameter to `sh` either. To fix this, I have now added a little bit
pre-processing to `$@` so that I can differentiate parameters before `--` and
parameters after `--`.